+++ public/react-services/pattern-handler.js
@@ -9,92 +9,60 @@
  *
  * Find more information about this on the LICENSE file.
  */
-import { GenericRequest } from './generic-request';
 import { AppState } from './app-state';
-import { WzMisc } from '../factories/misc';
 import { SavedObject } from './saved-objects';
 import { getDataPlugin, getToasts } from '../kibana-services';
 import { WazuhConfig } from '../react-services/wazuh-config';
+import { HEALTH_CHECK } from '../../common/constants';
 
 export class PatternHandler {
   /**
    * Get the available pattern list
    */
-  static async getPatternList(where) {
+  static async getPatternList(origin) {
     try {
-      var patternList = await SavedObject.getListOfWazuhValidIndexPatterns();
-
-      if (where === 'healthcheck') {
-        function getIndexPatterns() {
-          return new Promise(function (resolve, reject) {
-            setTimeout(async function () {
-              var patternList = await SavedObject.getListOfWazuhValidIndexPatterns();
-              resolve(patternList);
-            }, 500);
-          });
-        }
-        var i = 0;
-        // if the index pattern doesn't exist yet, we check 5 more times with a delay of 500ms
-        while (i < 5 && !patternList.length) {
-          i++;
-          patternList = await getIndexPatterns()
-            .then(
-              function (result) {
-                return result;
-              }
-            );
-        }
-      }
-
       const wazuhConfig = new WazuhConfig();
       const { pattern } = wazuhConfig.getConfig();
-      const indexPatternFound = patternList.find((indexPattern) => indexPattern.title === pattern);
 
-      if (!indexPatternFound) {
-        // if no valid index patterns are found we try to create the wazuh-alerts-*
-        try {
+      const defaultPatterns = [pattern];
+      const selectedPattern = AppState.getCurrentPattern();
+      if (selectedPattern && selectedPattern !== pattern) defaultPatterns.push(selectedPattern);
+      let patternList = await SavedObject.getListOfWazuhValidIndexPatterns(defaultPatterns, origin);
 
-          if (!pattern) return;
+      if (origin === HEALTH_CHECK) {
+        const indexPatternFound = patternList.find((indexPattern) => indexPattern.title === pattern);
 
-          getToasts().add({
-            color: 'warning',
-            title:
-              `No ${pattern} index pattern was found, proceeding to create it.`,
-            toastLifeTimeMs: 5000
-          });
+        if (!indexPatternFound && pattern) {
+          // if no valid index patterns are found we try to create the wazuh-alerts-*
+          try {
+            getToasts().add({
+              color: 'warning',
+              title:
+                `No ${pattern} index pattern was found, proceeding to create it.`,
+              toastLifeTimeMs: 5000
+            });
 
-          await SavedObject.createWazuhIndexPattern(pattern);
-          getToasts().addSuccess(`${pattern} index pattern created successfully`)
-          getDataPlugin().indexPatterns.setDefault(pattern, true);
-        } catch (err) {
-          getToasts().add({
-            color: 'error',
-            title: 'Error creating the index pattern.',
-            text: err.message || err,
-            toastLifeTimeMs: 3000
-          });
-          AppState.removeCurrentPattern();
+            if (await SavedObject.getExistingIndexPattern(pattern)) {
+              await SavedObject.refreshIndexPattern(pattern);
+              getToasts().addSuccess(`${pattern} index pattern updated successfully`);
+            } else {
+              await SavedObject.createWazuhIndexPattern(pattern);
+              getToasts().addSuccess(`${pattern} index pattern created successfully`);
+            }
 
-          this.wzMisc = new WzMisc();
-          this.wzMisc.setBlankScr(
-            'Sorry but no valid index patterns were found and creation was unsuccessful'
-          );
-          if (
-            !window.location.hash.includes('#/settings') &&
-            !window.location.hash.includes('#/blank-screen')
-          ) {
-            window.location.href = '/app/wazuh#/blank-screen/';
+            patternList = await SavedObject.getListOfWazuhValidIndexPatterns(defaultPatterns, origin);
+            !AppState.getCurrentPattern() && AppState.setCurrentPattern(pattern);
+          } catch (err) {
+            getToasts().addDanger({
+              title: 'Error creating the index pattern.',
+              text: err.message || err,
+              toastLifeTimeMs: 3000
+            });
+            AppState.removeCurrentPattern();
+
+            return;
           }
-          return;
         }
-        // retry again with the newly created index pattern
-        if (
-          !window.location.hash.includes('#/settings') &&
-          !window.location.hash.includes('#/health-check')
-        ) {
-          window.location.href = '/app/wazuh#/health-check/';
-        }
-        patternList = await SavedObject.getListOfWazuhValidIndexPatterns();
       }
       if (AppState.getCurrentPattern() && patternList.length) {
         let filtered = patternList.filter(
@@ -108,7 +76,6 @@ export class PatternHandler {
       console.error("getPatternList", error)
       throw new Error('Error Pattern Handler (getPatternList)');
     }
-    return;
   }
 
   /**